Bidirectional communication and streams for plugins
Plugins can now implement the `StreamingPlugin` trait, which allows them
to consume and produce raw `PipelineData`.

Both list streams and external streams are supported, and plugins can
handle both stdout and stderr simultaneously if they so wish.

This is not a breaking code change for Rust plugins, but it is
incompatible at the serialization level, so they do have to be
recompiled.

For plugins using the simpler `Plugin` trait, streams are still sent
as normal to the plugin, but they are internally collected to construct
a Value on the plugin side now instead of the engine.

This change lays the groundwork for bidirectional communication with
plugins in general. The API could be extended with functions to ask the
engine to evaluate objects, such as closures, that the plugin cannot,
though this is not a part of this patch.

It also allows for plugins to serve multiple calls within a single
process lifetime. While not currently being used, this could be helpful
in the future to make plugins execution cheaper to run in a loop, for
example by having the parser statically allocate the plugin executions
that will be necessary to run the command and reusing them for the
life of the pipeline.

Additionally, because this amounts to an almost complete protocol
overhaul anyway, a few extra changes were made in order to simplify
the protocol or provide a better user experience for future changes:

* Plugins don't start accepting input by default; they now require the
  `--stdio` argument. This frees us up to make changes to how plugins
  talk to nushell in the future by changing the command line arguments,
  and also gives users a less confusing experience if they try to run
  the binary without any args.

* Both the engine and plugins send a Hello message as a handshake at
  startup, and both verify the semver compatibility of their respective
  `nu-plugin` versions. An additional field is provided but unused for
  making semver-compatible optional protocol changes. Unrecognized
  values in this field are ignored intentionally. A version mismatch
  provides an appropriate error message to the user, rather than a
  vague failure to deserialize.

* The (de-)serialization of custom values has been improved to allow
  them to sit nested within a `Value`. The interfaces on either side are
  responsible for enforcing that the only custom value allowed to go
  through the plugin barrier is `PluginCustomValue`. The engine side
  ensures that plugins only get their own custom values just as before
  (with improved error messages), and the plugin side automatically
  serializes and deserializes just as before.

* As a consequence of the above: custom values are now supported as
  arguments, and embedded within streams. `PluginData` has been removed.

* Memory usage of `PluginCustomValue` is improved by keeping the
  identity of the plugin stored in a shared `Arc`, making it much more
  reasonable to send large volumes of custom values back and forth.

* `CollapseCustomValue` is renamed to the more general `ToBaseValue` as
  a variant of `CustomValueOp`. This will make it easier to support
  additional custom value ops in the future.

* `LabeledError` now uses a generic implementation based on
  `miette::Diagnostic` for `From<ShellError>`. This is more likely to
  produce good errors, and is important as it's more likely that plugins
  will have to handle a shell error embedded in a stream now. However,
  this is only a best effort, and errors containing multiple spans will
  have only one of their spans serialized.

The python example has been updated to reflect the changes, though
without demonstrating any stream functionality, and a new
`nu_plugin_stream_example` plugin has been added to demonstrate and
test stream functionality.