Add streaming API for plugins
Plugins can now implement the `StreamingPlugin` trait, which allows them
to consume and produce raw `PipelineData`.

Both list streams and external streams are supported, and plugins can
handle both stdout and stderr simultaneously if they so wish.

This is not a breaking code change for plugins, but it is incompatible
at the serialization level, so they do have to be recompiled.

For plugins using the simpler `Plugin` trait, streams are still sent
as normal to the plugin, but they are internally collected to construct
a Value on the plugin side now instead of the engine.

This change lays the groundwork for bidirectional communication with
plugins in general. The API could be extended with functions to ask the
engine to evaluate objects that the plugin cannot, though this is not
a part of this patch.

It also allows for plugins to serve multiple calls within a single
process lifetime. While not currently being used, this could be helpful
in the future to make plugins execution cheaper to run in a loop, for
example by having the parser statically allocate the plugin executions
that will be necessary to run the command and reusing them for the
life of the pipeline.