Add `tee` command for operating on copies of streams
This is inspired by the Unix tee command, but significantly more
powerful. Rather than just writing to a file, you can do any kind of
stream operation that Nushell supports within the closure.

The equivalent of Unix `tee -a file.txt` would be, for example,
`command | tee { save -a file.txt }` - but of course this is Nushell,
and you can do the same with structured data to JSON objects, or
even just run any other command on the system with it.

A `--stderr` flag is provided for operating on the stderr stream from
external programs. This may produce unexpected results if the stderr
stream is not then also printed by something else - nushell currently
doesn't.