Print stderr streams to stderr in `pipeline_data::print_if_stream()`
This corrects unexpected behavior if a stream from an external program
is transformed while still preserving its stderr output. Before this
change, that output is just drained and discarded. Worse, it's drained
to a buffer, which could be really slow and memory hungry if there's
a lot of output on stderr.

This is needed to make `tee --stderr` function in a non-surprising way.