Replace debug_assert! with assert! in Signature::check_names
Debug assertions don't run at release, which means that
`cargo test --release` fails because the tests for name checks don't run
properly. These checks are not really expensive, and there shouldn't
be any noticeable difference to startup time, so there isn't much reason
not to just leave them in.

It's valuable to be able to run `cargo test --release`, as that can
expose race conditions and dependencies on undefined behavior that
aren't exposed in debug builds.