Rename spans in the serialized form of Value
Span fields were previously renamed to `internal_span` to discourage
their use in Rust code, but this change also affected the serde I/O
for Value. I don't believe the Python plugin was ever updated to
reflect this change.

This effectively changes it back, but just for the serialized form.
There are good reasons for doing this:

1. `internal_span` is a much longer name, and would be one of the most
   common strings found in serialized Value data, probably bulking up
   the plugin I/O

2. This change was never really meant to have implications for plugins,
   and was just meant to be a hint that `.span()` should be used
   instead in Rust code.

When Span refactoring is complete, the serialized form of Value will
probably change again in some significant way, so I think for now it's
best that it's left like this.

This has implications for #11911, particularly for documentation and
for the Python plugin as that was already updated in that PR to
reflect `internal_span`. If this is merged first, I will update that PR.

This would probably be considered a breaking change as it would break
plugin I/O compatibility (but not Rust code). I think it can probably go
in any major release though - all things considered, it's pretty minor,
and users are already expected to recompile plugins for new major
versions. However, it may also be worth holding off to do it together
with #11911 as that PR makes breaking changes in general a little bit
friendlier.