Plugin StreamReader: fuse the iterator after an error
This patches `StreamReader`'s iterator implementation to not return any
values after an I/O error has been encountered.

Without this, it's possible for a protocol error to cause the channel
to disconnect, in which case every call to `recv()` returns an error,
which causes the iterator to produce error values infinitely. There
are some commands that don't immediately stop after receiving an error
so it's possible that they just get stuck in an infinite error. This
fixes that so the error is only produced once, and then the stream ends
artificially.