Remove feature `extra`
The effect has been undermined by introducing the full builds on our
release pages and having more activity on some of the extra commands.

To simplify the feature matrix let's get rid of it and focus our effort
on truly either refining a command to well-specified behavior or
discarding it entirely from the `nu` binary and moving it into plugins.