Disable `fmt` feature of `polars(-core)`
In normal operations we don't display the dataframes directly.
The `fmt` feature on `polars-core` pulls in the `comfy-table` crate with
its own dependencies on `crossterm` and `strum(_macros)`. This has the
chance to duplicate dependencies. (currently strum version divergence)

Without this feature only the shapes should be displayed.

May degrade the error output during testing.