Make the plugin persistence GC delay test more reliable
This makes the test a bit more complicated, but implements a timeout
loop in the script. As long as the test completes in 5 seconds it's
considered to be ok. The default is 10 seconds, so that would still
be half that.

This should help with running on the busy CI where things sometimes take
a while. Unfortunately this is a timing sensitive test. The alternative
is basically to just not test this at all because it's too difficult
to guarantee that it will complete in time. If we continue to have
issues, I might just have to take that route instead.