Add environment engine calls for plugins
This adds two engine calls: `GetEnvVar` and `GetEnvVars`, for getting
environment variables from the plugin command context.

Additionally, two convenience methods are provided for using the
correct working directory for the command if required.
`EngineInterface::get_current_dir` gets the current directory from the
`PWD` environment variable, and `EngineInterface::with_current_dir` sets
it as the thread current directory for the duration of a function.

Plugins are now launched in the directory of their executable to try to
make improper use of the current directory without first setting it
more obvious. Plugins previously launched in whatever the current
directory of the engine was at the time the plugin command was run,
but switching to persistent plugins broke this, because they stay in
whatever directory they launched in initially.

This also fixes the `gstat` plugin to use `with_current_dir()`, which
was directly affected by this problem.